revise model's constitutive relation(createdPosts/savedPosts)
