+++ src/server/users/createdPostsModel.js
@@ -1,0 +1,21 @@
+import db from '../../server/database';
+import {KeyUtils, ENTITY} from '../../server/key-utils';
+
+export default class CreatedPostManager {
+  static addCreatedPost(entity, entityKey, userKey, timeHash, cb) {
+    if (entity !== ENTITY.ITEM && entity !== ENTITY.IMAGE) {
+      return cb(new Error(('invalid entity'), null));
+    }
+    const idxKey = KeyUtils.getIdxKey(ENTITY.CREATED_POST, timeHash, userKey);
+    const idxPost = {
+      entity: entity,
+      key: entityKey
+    };
+    return db.put(idxKey, idxPost, (err) => {
+      if (err) {
+        return cb(new Error('error while putting in DB'), null);
+      }
+      return cb(null, idxKey);
+    });
+  }
+}
