+++ src/test/node/index.js
@@ -11,9 +11,9 @@ import './items/models';
 import './images/models';
 import './images/controllers';
 import './users/controllers';
-import './PostManager';
+import './users/createdPostsModel';
+import './users/savedPostsModel';
 import './database';
 import './aws-s3';
 import './auth-token';
 import './items/models';
-import './users/controllers';
